+++ packages/components/components/elvis-box/src/react/elvia-box.tsx
@@ -1,14 +1,24 @@
-import React, { FC, useEffect, useRef } from 'react';
+import React, { CSSProperties, FC, useEffect, useRef } from 'react';
 import { BoxArea, BoxColoredLine, BoxTitle, BoxContent } from './styledComponents';
 export interface BoxProps {
   content: string | HTMLElement;
   title?: string;
   isColored?: boolean;
   hasBorder?: boolean;
+  className?: string;
+  inlineStyle?: { [style: string]: CSSProperties };
   webcomponent: any;
 }
 
-const Box: FC<BoxProps> = ({ content, title, isColored = false, hasBorder = false, webcomponent }) => {
+const Box: FC<BoxProps> = ({
+  content,
+  title,
+  isColored = false,
+  hasBorder = false,
+  className,
+  inlineStyle,
+  webcomponent,
+}) => {
   const boxContent = useRef<HTMLDivElement>(null);
   const boxTitle = useRef<HTMLDivElement>(null);
 
@@ -29,22 +39,24 @@ const Box: FC<BoxProps> = ({ content, title, isColored = false, hasBorder = fals
   }, [webcomponent]);
 
   return (
-    <BoxArea>
-      {title && <BoxTitle data-testid="box-title">{title}</BoxTitle>}
-      {!title && <BoxTitle ref={boxTitle}></BoxTitle>}
-      {content && (
-        <BoxContent hasBorder={hasBorder} data-testid="box-content">
-          {isColored && <BoxColoredLine data-testid="box-colored-line"></BoxColoredLine>}
-          {content}
-        </BoxContent>
-      )}
-      {!content && (
-        <BoxContent hasBorder={hasBorder} data-testid="box-no-content">
-          {isColored && <BoxColoredLine></BoxColoredLine>}
-          <div ref={boxContent}></div>
-        </BoxContent>
-      )}
-    </BoxArea>
+    <div className={`${className ? className : ''}`} style={inlineStyle}>
+      <BoxArea data-testid="box-area">
+        {title && <BoxTitle data-testid="box-title">{title}</BoxTitle>}
+        {!title && <BoxTitle ref={boxTitle}></BoxTitle>}
+        {content && (
+          <BoxContent hasBorder={hasBorder} data-testid="box-content">
+            {isColored && <BoxColoredLine data-testid="box-colored-line"></BoxColoredLine>}
+            {content}
+          </BoxContent>
+        )}
+        {!content && (
+          <BoxContent hasBorder={hasBorder} data-testid="box-no-content">
+            {isColored && <BoxColoredLine></BoxColoredLine>}
+            <div ref={boxContent}></div>
+          </BoxContent>
+        )}
+      </BoxArea>
+    </div>
   );
 };