+++ packages/components/components/elvis-progress-linear/src/react/elvia-progress-linear.test.js
@@ -4,6 +4,7 @@ import { mount } from 'enzyme';
 
 describe('Elvis Progress Linear', () => {
   let wrapper;
+  let progressWrapper;
   let progressLinear;
 
   describe('Default', () => {
@@ -47,6 +48,19 @@ describe('Elvis Progress Linear', () => {
     });
     it('should have class ewc-progress-linear--error', function (done) {
       expect(progressLinear.getDOMNode()).toHaveClass('ewc-progress-linear--error');
+      done();
+    });
+  });
+  describe('className and inlineStyle passed to wrapper', () => {
+    beforeEach(() => {
+      wrapper = mount(
+        <ProgressLinear value="48" className="test-class" inlineStyle={{ margin: '24px' }}></ProgressLinear>,
+      );
+      progressWrapper = wrapper.find({ 'data-testid': 'progress-wrapper' }).at(0);
+    });
+    it('should have className and inlineStyle', function (done) {
+      expect(progressWrapper.getDOMNode()).toHaveStyle('margin: 24px');
+      expect(progressWrapper.getDOMNode()).toHaveClass('test-class');
       done();
     });
   });