+++ packages/components/components/elvis-radio-filter/src/react/elvia-radio-filter.test.js
@@ -43,4 +43,18 @@ describe('Elvis RadioFilter', () => {
       done();
     });
   });
+  describe('className and inlineStyle passed to wrapper', () => {
+    beforeEach(() => {
+      wrapper = mount(<RadioFilter items={items} className="test-class" inlineStyle={{ margin: '24px' }} />);
+      radioFilterElements = wrapper.find({ 'data-testid': 'radio-filter-group' }).at(0);
+    });
+    afterEach(() => {
+      wrapper.unmount();
+    });
+    it('should have className and inlineStyle', function (done) {
+      expect(radioFilterElements.getDOMNode()).toHaveStyle('margin: 24px');
+      expect(radioFilterElements.getDOMNode()).toHaveClass('test-class');
+      done();
+    });
+  });
 });