+++ packages/components/components/elvis-tabs/src/react/elvia-tabs.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useEffect, useState, useRef } from 'react';
+import React, { FC, useEffect, useState, useRef, CSSProperties } from 'react';
 import classNames from 'classnames';
 import toolbox from '@elvia/elvis-toolbox';
 import './style.scss';
@@ -7,10 +7,20 @@ export interface TabsProps {
   value: number;
   isInverted: boolean;
   valueOnChange?: (value: number) => void;
+  className?: string;
+  inlineStyle?: { [style: string]: CSSProperties };
   webcomponent?: any;
 }
 
-const Tabs: FC<TabsProps> = ({ items, value = 0, isInverted, valueOnChange, webcomponent }) => {
+const Tabs: FC<TabsProps> = ({
+  items,
+  value = 0,
+  isInverted,
+  valueOnChange,
+  className,
+  inlineStyle,
+  webcomponent,
+}) => {
   const [currValue, setCurrValue] = useState(value);
   const [isOnRightEnd, setIsOnRightEnd] = useState(true);
   const [isOnLeftEnd, setIsOnLeftEnd] = useState(true);
@@ -116,7 +126,12 @@ const Tabs: FC<TabsProps> = ({ items, value = 0, isInverted, valueOnChange, webc
   });
 
   return (
-    <div className={tabsClasses} ref={tabsRef} data-testid="tabs-container">
+    <div
+      className={tabsClasses + (className ? ' ' + className : '')}
+      style={inlineStyle}
+      ref={tabsRef}
+      data-testid="tabs-container"
+    >
       <div
         className={arrowLeftClasses}
         onClick={() => {