+++ packages/components/components/elvis-chip/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Elvia Chip Changelog
 
+## 1.2.0 (07.02.22)
+
+### New feature
+
+- Added inlineStyle and className props
+
 ## 1.1.7 (26.01.21)
 
 - Updated dependencies